Split kong tests to a separate file
